hotfix and separate code with bugs into new file
